requirements.txt
emoji==2.12.1
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.36.0
streamlit_lottie==0.0.5
